% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootpval.R
\name{bootpval}
\alias{bootpval}
\title{Bootstrap P-Values}
\usage{
bootpval(x, conf.level = 0.95, iter = 3000, mu0 = 0, test = "two")
}
\arguments{
\item{x}{vector of data}

\item{conf.level}{number between 0 and 1 (1 - alpha)}

\item{iter}{number; integer of bootstrap iterations}

\item{mu0}{number; null hypothesis for mu}

\item{test}{string; "two", "upper", or "lower", indicating test tails}
}
\value{
list; p-value, tcalc, n, x, test, and conf.int
}
\description{
Bootstrap P-Values
}
\examples{
x=rnorm(30); bootpval(x)
}
