% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{Bootstrap Sampling}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{number; the number of samples to take from the data provided}

\item{x}{vector; a vector containing the data to sample from}

\item{fun}{string; the function name to apply across each sample, calculating the sample statistic}

\item{alpha}{number; a number between 0 and 1 indicating the confidence interval}

\item{cx}{number; graphical parameter}

\item{...}{additional graphing parameters for hist()}
}
\value{
list containg the confidence interval, function, data, and histogram parameters
}
\description{
Bootstrap Sampling
}
\examples{
data=rnorm(1:100); myboot2(x = data, alpha=0.05, iter=10000, fun="mean", col="red")
}
