% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltp.R
\name{mycltp}
\alias{mycltp}
\title{Poisson CLT}
\usage{
mycltp(n, iter, lambda = 10, ...)
}
\arguments{
\item{n}{number; sample size}

\item{iter}{number; number of samples to draw}

\item{lambda}{number; lambda parameter for the Poisson distribution}

\item{...}{histogram plotting parameters}
}
\value{
series of three plots
}
\description{
Poisson CLT
}
\examples{
mycltp(30, 10000)
}
